fix bugs & modulation form modify responsive css
