+++ client/src/Components/NavBar/NavBar.module.css
@@ -2,7 +2,8 @@
   width: 100%;
   height: 100px; /* increase height to provide more space */
   display: flex;
-  position: fixed;
+  position: sticky;
+  top: 0;
   z-index: 3;
   backdrop-filter: blur(5px);
   align-items: center;
