+++ client/src/Views/Form/Form.jsx
@@ -1,117 +1,8 @@
-import { useDispatch } from 'react-redux';
-import { useState } from 'react';
-import { validateError, validate, validateColorError } from '../../Components/formValidation/errorHandler';
-import {postClothes} from '../../Redux/ActionsGet';
-import uploadImage from './uploadImage';
+import { useForm } from './formHook'
+
 
 function Form() {
-	const dispatch = useDispatch();
-
-	const [form, setForm] = useState({
-		name: "",
-		size: "",
-		price: "",
-		type: "",
-		image: "",
-		sex: [],
-		stockGeneral: 0,
-		stockSize: 0,
-		colors: [],
-		existing: true,
-	});
-	const [addColor, setAddColor] = useState({
-		color: '',
-		stockColors: 0,
-	});
-  ///////////////// Validation ///////////////////////////
-	const [error, setError] = useState({});
-	const [validated, setValidated] = useState({});
-	const [colorError, setColorError] = useState({});
-  //////////////////////////////////////////////
-	const handleForm = async (event) => {
-		if (event.target.name === 'image') {
-			setError(validateError({ ...form, [event.target.name]: await uploadImage(event.target.files[0]) }));
-			setValidated(validate({ ...form, [event.target.name]: await uploadImage(event.target.files[0]) }));
-			setForm({
-				...form,
-				[event.target.name]: await uploadImage(event.target.files[0]),
-			});
-		} else {
-			setError(validateError({ ...form, [event.target.name]: event.target.value }));
-			setValidated(validate({ ...form, [event.target.name]: event.target.value }));
-			setForm({...form, [event.target.name]: event.target.value});
-		}
-	};
-
-	const genreHandler = (e) => {
-		let value = e.target.value;
-		if (value === "") alert("Please select a genre");
-		else if (form.sex?.includes(value))
-			alert("The genre has already been selected");
-		else {
-			setError(validateError({ ...form, sex: value }));
-			setValidated(validate({ ...form, sex: value }));
-				setForm((prevState) => ({
-				...prevState,
-				sex: [...prevState.sex, value],
-			}));
-		}
-	};
-	const genreDeleteHandler = (e) => {
-		const value = e.target.value;
-		const filteredGenre = form.sex?.filter((genre) => genre !== value.toString());
-		setError(validateError({ ...form, sex: filteredGenre }));
-		setValidated(validate({ ...form, sex: filteredGenre }));
-		setForm({ ...form, sex: filteredGenre });
-	};
-
-	const handlerColor = (e) => {
-		const value = e.target.value;
-		const event = e.target.name;
-
-		if(value === 'color') value.toLowerCase();
-		setColorError(validateColorError({...addColor, [event] : value }));
-		setAddColor({...addColor, [event] : value });
-	};
-	const handlerAddColor = (e) => {
-		e.preventDefault();
-		setError(validateError({ ...form, colors: [...form.colors, addColor] }));
-		setValidated(validate({ ...form, colors: [...form.colors, addColor] }));
-		setForm((prevState) => ({
-				...prevState,
-				colors: [...form.colors, addColor],
-			}));
-	};
-	const colorDeleteHandler = (e) => {
-		const key = e.target.key;
-		const index = form.colors?.findIndex((color) => color.key === key);
-		if (index !== -1) {
-			const filteredColors = [...form.colors];
-			filteredColors.splice(index, 1);
-			setError(validateError({ ...form, colors: filteredColors }));
-			setValidated(validate({ ...form, colors: filteredColors }));
-			setForm({ ...form, colors: filteredColors });
-		}
-	};
-
-	function SubmitHandler(e) {
-		e.preventDefault();
-		dispatch(postClothes(form));
-		setForm({
-			name: '',
-			size: '',
-			price: '',
-			type: '',
-			image: '',
-			sex: [],
-			stockGeneral: 0,
-			stockSize: 0,
-			colors: [],
-			existing: true,
-		});
-		alert('The cloth has been submitted');
-	}
-
+	const { form,  addColor,  error,  validated,  colorError,  handleForm, genreHandler, genreDeleteHandler, handlerColor, handlerAddColor, colorDeleteHandler, SubmitHandler } = useForm();
 	let styleError = {
 		fontWeight: 'bold',
 		color: '#dc3545',
@@ -121,6 +12,8 @@ function Form() {
 		fontWeight: 'bold',
 		color: ' #22a922',
 	};
+
+	console.log(form.image);
 
 	return (
