+++ client/src/Views/Home/Home.module.css
@@ -1,4 +1,8 @@
-.container{
-
-
+.img {
+    position: absolute;
+    top: 0;
+    width: 100vw;
+    height: 100vh;
+    object-fit: cover;
+    z-index: -10;
 }
